Render birth form
