+++ src/features/config/form/index.ts
@@ -13,8 +13,7 @@
 import { birthRegisterForms as birth } from './birth'
 import { deathRegisterForms as death } from './death'
 import { marriageRegisterForms as marriage } from './marriage'
-import { populateRegisterFormsWithAddresses } from './addresses'
-import { IForms, Event } from './types'
+import { IForms } from './types'
 
 export async function formHandler(): Promise<IForms> {
   // AS OF OPENCRVS v1.3.0 YOU CANNOT MAKE A CHANGE TO A CIVIL REGISTRATION FORM ONCE YOU GO LIVE IN PRODUCTION
@@ -41,8 +40,8 @@ export async function formHandler(): Promise<IForms> {
     // YOU CAN NAME THIS VERSION ANYTHING YOU LIKE AS LONG AS IT IS UNIQUE EACH TIME.
     // WE RECOMMEND USING SEMANTIC VERSION NUMBERING
     version: 'v1.0.0',
-    birth: populateRegisterFormsWithAddresses(birth, Event.Birth),
-    death: populateRegisterFormsWithAddresses(death, Event.Death),
-    marriage: populateRegisterFormsWithAddresses(marriage, Event.Marriage)
+    birth,
+    death,
+    marriage
   }
 }
